fix: extract LNURL session into app-level provider
Move the SSE connection lifecycle from the useLnurlSession hook into
a new LnurlProvider so the session survives component unmount. When
the user navigates away from the receive screen, the LNURL remains
active and can still serve invoice requests from payers.
