+++ src/index.tsx
@@ -15,6 +15,7 @@ import { LimitsProvider } from './providers/limits'
 import { NudgeProvider } from './providers/nudge'
 import * as Sentry from '@sentry/react'
 import { SwapsProvider } from './providers/swaps'
+import { LnurlProvider } from './providers/lnurl'
 import { shouldInitializeSentry } from './lib/sentry'
 import { FeesProvider } from './providers/fees'
 import { AnnouncementProvider } from './providers/announcements'
@@ -75,21 +76,23 @@ root.render(
             <FlowProvider>
               <WalletProvider>
                 <SwapsProvider>
-                  <LimitsProvider>
-                    <FeesProvider>
-                      <OptionsProvider>
-                        <NudgeProvider>
-                          <AnnouncementProvider>
-                            <ToastProvider>
-                              <ErrorBoundary>
-                                <App />
-                              </ErrorBoundary>
-                            </ToastProvider>
-                          </AnnouncementProvider>
-                        </NudgeProvider>
-                      </OptionsProvider>
-                    </FeesProvider>
-                  </LimitsProvider>
+                  <LnurlProvider>
+                    <LimitsProvider>
+                      <FeesProvider>
+                        <OptionsProvider>
+                          <NudgeProvider>
+                            <AnnouncementProvider>
+                              <ToastProvider>
+                                <ErrorBoundary>
+                                  <App />
+                                </ErrorBoundary>
+                              </ToastProvider>
+                            </AnnouncementProvider>
+                          </NudgeProvider>
+                        </OptionsProvider>
+                      </FeesProvider>
+                    </LimitsProvider>
+                  </LnurlProvider>
                 </SwapsProvider>
               </WalletProvider>
             </FlowProvider>
